+++ src/ip_info.hpp
@@ -13,7 +13,7 @@
 class IpInfo {
     static bool initialized;
 
-    static std::map<uint32_t, std::string> asNames;
+    static std::map<int64_t, std::string> asNames;
     static PrefixContainer<int64_t> asNumbers;
     static PrefixContainer<std::string> ixNames;
 
@@ -21,25 +21,25 @@ class IpInfo {
     void initializeAsToOwner();
     void initializeIpToIx();
 
-    public:
-        static const int64_t UnknownAsn;
+public:
+    static const int64_t UnknownAsn;
 
-        IpInfo();
-        
-        /**
-         * @returns AS number for the given address or AsNumber::UnknownAsn if not found.
-        */
-        int64_t getAsNumber(Ipv4Address ipv4_address);
+    IpInfo();
+    
+    /**
+     * @returns AS number for the given address or AsNumber::UnknownAsn if not found.
+    */
+    int64_t getAsNumber(Ipv4Address ipv4_address);
 
-        /**
-         * @returns short string describing the owner of the AS
-        */
-        std::string getAsName(int64_t asn);
+    /**
+     * @returns short string describing the owner of the AS
+    */
+    std::string getAsName(int64_t asn);
 
-        /**
-         * @returns short string naming the IXP or empty if address is not IX peering router
-        */
-        std::string getIx(Ipv4Address addr);
+    /**
+     * @returns short string naming the IXP or empty if address is not IX peering router
+    */
+    std::string getIx(Ipv4Address addr);
 };
 
 #endif
